sparse-index: make non-mixed `reset` sparse-aware
This change allows non-mixed resets to no longer require that the full index
is expanded. This is accomplished by updating the `prime_cache_tree_rec`
function to have it reconstruct the cache tree "aware" of whether a directory
is sparse in the index or not. This check must be done by verifying the
contents of the cache itself (rather than checking whether a directory is
inside of the sparse checkout cone) as entries may have been added to the
sparse index outside of the checkout cone.

Signed-off-by: Victoria Dye <vdye@github.com>