sparse index: enable only for git repos
Check whether git dir exists before adding any repo settings. If it
does not exist, BUG with the message that one cannot add settings for an
uninitialized repository. If it does exist, proceed with adding repo
settings.

Additionally, ensure the above BUG is not triggered in the pack-objects
and checkout builtins when users pass the -h flag by moving existing calls
to prepare_repo_settings() to follow parsing of this option. Ensure it is
not triggered for commit-graph by returning early if the git dir does not
exist and for test-read-cache by moving repo setup to occur after setup of
repo and git dir.

Signed-off-by: Lessley Dennington <lessleydennington@gmail.com>