build-git-installers: update windows signing
This commit contains two temporary workflows:

1. windows-full: This workflow was designed to use the same key used for Linux
signing for GPG signing. It also uses the Azure Code Signing action to sign
Windows installers.
2. windows-minimal: This workflow was intended to test the minimal requirements
for Windows builds to see whether it is possible to use Azure Code Signing as
the singtool alias for exe signing. This workflow is already set up with tmate
for testing on GitHub Actions runners (since the issue described below does not
repro locally).

While we were unable to get past an issue with the second workflow hanging
instead of executing signing, hopefully this will provide a solid starting
point if someone would like to re-attempt Azure Code Signing/GPG signing for
microsoft/git in the future. Note: If signing is successfully implemented with
the alias used in workflow #2, this alias should also be used for installer
signing rather than the Azure Code Signing action so that the embedded
uninstaller is signed in addition to the installers.